* 0j0j0j
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@1352 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ player.cpp
@@ -1393,7 +1393,7 @@ void Player::onCreatureAppear(const Creature* creature, bool isLogin)
 		{
 			int32_t period = (int32_t)time(NULL) - lastLogout - 600;
 			if(period > 0)
-				useStamina(-(std::min(getSpentStamina(), (uint64_t)std::abs((uint64_t)period * g_config.getNumber(
+				useStamina(-int64_t(std::min(getSpentStamina(), (uint64_t)std::abs((uint64_t)period * g_config.getNumber(
 					ConfigManager::RATE_STAMINA_GAIN)))));
 		}
